[MPDF-56] Pdf generation fails when there is no site
git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk/maven-pdf-plugin@1344160 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/maven/plugins/pdf/PdfMojo.java
@@ -592,7 +592,11 @@ public class PdfMojo
         tmpSiteDir.mkdirs();
 
         // copy site
-        FileUtils.copyDirectoryStructure( siteDirectory, tmpSiteDir );
+        if ( siteDirectory.exists() )
+        {
+            FileUtils.copyDirectoryStructure( siteDirectory, tmpSiteDir );
+        }
+
         // Remove SCM files
         List files =
             FileUtils.getFileAndDirectoryNames( tmpSiteDir, FileUtils.getDefaultExcludesAsString(), null, true,
